@@ -2,9 +2,15 @@
   "name": "cucumber-expressions",
   "version": "1.0.1",
   "description": "Cucumber Expressions - a simpler alternative to Regular Expressions",
+  "main": "dist/index.js",
   "scripts": {
-    "test": "mocha --recursive",
-    "coverage": "ISTANBUL=true istanbul cover _mocha -- --recursive",
+    "test": "npm run eslint && npm run coverage",
+    "eslint": "eslint src test",
+    "build": "babel src --out-dir dist",
+    "prepublish": "npm run build",
+    "postinstall": "node scripts/postinstall.js",
+    "mocha": "mocha",
+    "coverage": "ISTANBUL=true istanbul cover _mocha",
     "postcoverage": "istanbul check-coverage --statements 100 --branches 100"
   },
   "repository": {
@@ -24,8 +30,12 @@
   },
   "homepage": "https://github.com/cucumber/cucumber-expressions-javascript#readme",
   "devDependencies": {
-    "eslint": "^2.13.1",
+    "babel-cli": "^6.11.4",
+    "babel-preset-es2015": "^6.13.2",
+    "eslint": "^3.5.0",
+    "eslint-config-eslint": "^3.0.0",
+    "eslint-plugin-node": "^2.0.0",
     "istanbul": "^0.4.0",
-    "mocha": "^2.3.3"
+    "mocha": "^3.0.2"
   }
 }
